requirements.txt
Django==1.11.7
django-filter==1.1.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
Markdown==2.6.9
PyJWT==1.5.3
pytz==2017.3
